requirements.txt
google-cloud-aiplatform
google-cloud-firestore
streamlit
pandas
python-dotenv